prepare-data: fix LAU match bug in IT/1991
IT/1991 showed missing data due to the LAUs in tha data having 5
digits instead of 6. This was due to mixed dtypes in the lau
column (one lau being . ), which led to the lau transformation code
not running as it only operated on integer colums.

Read CSV in with dtype=str and do transformation separately.

Also add -p, --prefix flag to only process data for certain
countries.